+++ frontend/upyog-ui/web/micro-ui-internals/packages/modules/ndc/src/pages/employee/Inbox/SearchFormFieldsComponent.js
@@ -10,9 +10,9 @@ const SearchFormFieldsComponents = ({ registerRef, searchFormState, searchFieldC
   if (!isMobile) {
     return (
       <React.Fragment>
-        <div className="search-container ndc-search-container-custom" >
+        <div className="search-container" style={{width: "200%",marginLeft: "24px",}}>
           <div className="search-complaint-container">
-            <div className="complaint-input-container ndc-complaint-input-container" >
+            <div className="complaint-input-container" style={{ textAlign: "start" }} >
               <SearchField>
                 <label>{t("NOC_HOME_SEARCH_RESULTS_APP_NO_LABEL")}</label>
                 <TextInput name="applicationNo" inputRef={registerRef("applicationNo").ref} onChange={registerRef("applicationNo").onChange} onBlur={registerRef("applicationNo").onBlur}/>
@@ -21,7 +21,7 @@ const SearchFormFieldsComponents = ({ registerRef, searchFormState, searchFieldC
                 <label>{t("CORE_COMMON_PHONE_NUMBER")}</label>
                 <TextInput name="mobileNumber" inputRef={registerRef("mobileNumber").ref} onChange={registerRef("mobileNumber").onChange} onBlur={registerRef("mobileNumber").onBlur}/>
               </SearchField>
-              <div className="search-action-wrapper ndc-search-action-wrapper" >
+              <div className="search-action-wrapper" style = {{width: "100%"}}>
                 {searchFieldComponents}
               </div>
             </div>