+++ upyog-automation-testing/src/main/java/org/upyog/Automation/Modules/CnD/CndVendor.java
@@ -21,11 +21,11 @@ public class CndVendor {
             TestEngine engine =
                     new TestEngine(
                             driver,
-                            "src/main/resources/config/dev.properties"
+                            "config/dev.properties"
                     );
 
             engine.executeModule(
-                    "src/main/resources/test-config/cnd/cnd_vendor_module.json"
+                    "test-config/cnd/cnd_vendor_module.json"
             );
 
         } catch (Exception e) {