+++ upyog-automation-testing/src/main/java/org/upyog/Automation/Modules/OBPAS/OBPASEmp.java
@@ -21,11 +21,11 @@ public class OBPASEmp {
             TestEngine engine =
                     new TestEngine(
                             driver,
-                            "src/main/resources/config/dev.properties"
+                            "config/dev.properties"
                     );
 
             engine.executeModule(
-                    "src/main/resources/test-config/obpas/obpas_employee_module.json"
+                    "test-config/obpas/obpas_employee_module.json"
             );
 
         }catch(Exception e){