+++ upyog-automation-testing/src/main/java/org/upyog/Automation/Modules/WaterAndSewerage/SewerageEmp.java
@@ -19,11 +19,11 @@ public class SewerageEmp {
             TestEngine engine =
                     new TestEngine(
                             driver,
-                            "src/main/resources/config/dev.properties"
+                            "config/dev.properties"
                     );
 
             engine.executeModule(
-                    "src/main/resources/test-config/waterandsewerage/sewerage_employee_module.json"
+                    "test-config/waterAndSewerage/sewerage_employee_module.json"
             );
 
         }catch(Exception e){