+++ upyog-automation-testing/src/main/java/org/upyog/Automation/model/ModuleConfig.java
@@ -1,4 +1,0 @@
-package org.upyog.Automation.model;
-
-public class ModuleConfig {
-}